+++ src/draw_target.rs
@@ -3,18 +3,23 @@ use embedded_graphics::{
     pixelcolor::BinaryColor, primitives::Rectangle, Pixel,
 };
 
-pub struct MonoFontDrawTarget<'a, T, C> {
+pub struct MultiMonoFontDrawTarget<'a, T, C> {
     parent: &'a mut T,
-    colors: C,
+    text_color: C,
+    background_color: Option<C>,
 }
 
-impl<'a, T: DrawTarget, C> MonoFontDrawTarget<'a, T, C> {
-    pub fn new(parent: &'a mut T, colors: C) -> Self {
-        Self { parent, colors }
+impl<'a, T: DrawTarget, C> MultiMonoFontDrawTarget<'a, T, C> {
+    pub fn new(parent: &'a mut T, text_color: C, background_color: Option<C>) -> Self {
+        Self {
+            parent,
+            text_color,
+            background_color,
+        }
     }
 }
 
-impl<T: DrawTarget> DrawTarget for MonoFontDrawTarget<'_, T, Foreground<T::Color>> {
+impl<T: DrawTarget> DrawTarget for MultiMonoFontDrawTarget<'_, T, T::Color> {
     type Color = BinaryColor;
     type Error = T::Error;
 
@@ -22,14 +27,23 @@ impl<T: DrawTarget> DrawTarget for MonoFontDrawTarget<'_, T, Foreground<T::Color
     where
         I: IntoIterator<Item = Self::Color>,
     {
-        let foreground_color = self.colors.0;
-
         self.parent.draw_iter(
             colors
                 .into_iter()
                 .into_pixels(area)
-                .filter(|Pixel(_, color)| color.is_on())
-                .map(|Pixel(pos, _)| Pixel(pos, foreground_color)),
+                .filter(|Pixel(_, color)| color.is_on() || self.background_color.is_some())
+                .map(|Pixel(pos, pixel_color)| {
+                    let color = if pixel_color.is_on() {
+                        self.text_color
+                    } else {
+                        if let Some(background_color) = self.background_color {
+                            background_color
+                        } else {
+                            self.text_color
+                        }
+                    };
+                    Pixel(pos, color)
+                }),
         )
     }
 
@@ -42,8 +56,14 @@ impl<T: DrawTarget> DrawTarget for MonoFontDrawTarget<'_, T, Foreground<T::Color
 
     fn fill_solid(&mut self, area: &Rectangle, color: Self::Color) -> Result<(), Self::Error> {
         match color {
-            BinaryColor::On => self.parent.fill_solid(area, self.colors.0),
-            BinaryColor::Off => Ok(()),
+            BinaryColor::On => self.parent.fill_solid(area, self.text_color),
+            BinaryColor::Off => {
+                if let Some(background_color) = self.background_color {
+                    self.parent.fill_solid(area, background_color)
+                } else {
+                    Ok(())
+                }
+            }
         }
     }
 
@@ -52,89 +72,8 @@ impl<T: DrawTarget> DrawTarget for MonoFontDrawTarget<'_, T, Foreground<T::Color
     }
 }
 
-impl<T: DrawTarget> DrawTarget for MonoFontDrawTarget<'_, T, Background<T::Color>> {
-    type Color = BinaryColor;
-    type Error = T::Error;
-
-    fn fill_contiguous<I>(&mut self, area: &Rectangle, colors: I) -> Result<(), Self::Error>
-    where
-        I: IntoIterator<Item = Self::Color>,
-    {
-        let foreground_color = self.colors.0;
-
-        self.parent.draw_iter(
-            colors
-                .into_iter()
-                .into_pixels(area)
-                .filter(|Pixel(_, color)| color.is_off())
-                .map(|Pixel(pos, _)| Pixel(pos, foreground_color)),
-        )
-    }
-
-    fn draw_iter<I>(&mut self, _pixels: I) -> Result<(), Self::Error>
-    where
-        I: IntoIterator<Item = Pixel<Self::Color>>,
-    {
-        unreachable!()
-    }
-
-    fn fill_solid(&mut self, area: &Rectangle, color: Self::Color) -> Result<(), Self::Error> {
-        match color {
-            BinaryColor::On => Ok(()),
-            BinaryColor::Off => self.parent.fill_solid(area, self.colors.0),
-        }
-    }
-
-    fn clear(&mut self, _color: Self::Color) -> Result<(), Self::Error> {
-        unreachable!()
-    }
-}
-
-impl<T: DrawTarget> DrawTarget for MonoFontDrawTarget<'_, T, Both<T::Color>> {
-    type Color = BinaryColor;
-    type Error = T::Error;
-
-    fn fill_contiguous<I>(&mut self, area: &Rectangle, colors: I) -> Result<(), Self::Error>
-    where
-        I: IntoIterator<Item = Self::Color>,
-    {
-        let foreground_color = self.colors.0;
-        let background_color = self.colors.1;
-
-        self.parent.fill_contiguous(
-            area,
-            colors.into_iter().map(|color| match color {
-                BinaryColor::Off => background_color,
-                BinaryColor::On => foreground_color,
-            }),
-        )
-    }
-
-    fn draw_iter<I>(&mut self, _pixels: I) -> Result<(), Self::Error>
-    where
-        I: IntoIterator<Item = Pixel<Self::Color>>,
-    {
-        unreachable!()
-    }
-
-    fn fill_solid(&mut self, area: &Rectangle, color: Self::Color) -> Result<(), Self::Error> {
-        match color {
-            BinaryColor::On => self.parent.fill_solid(area, self.colors.0),
-            BinaryColor::Off => self.parent.fill_solid(area, self.colors.1),
-        }
-    }
-
-    fn clear(&mut self, _color: Self::Color) -> Result<(), Self::Error> {
-        unreachable!()
-    }
-}
-
-impl<T: DrawTarget, C> Dimensions for MonoFontDrawTarget<'_, T, C> {
+impl<T: DrawTarget, C> Dimensions for MultiMonoFontDrawTarget<'_, T, C> {
     fn bounding_box(&self) -> Rectangle {
         self.parent.bounding_box()
     }
 }
-
-pub struct Foreground<C>(pub C);
-pub struct Background<C>(pub C);
-pub struct Both<C>(pub C, pub C);