+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 0.3.0 - 2024-09-27
+- optimize StaticText multi-line text
+- add no_std attribute
+- optimize MultiMonoFontDrawTarget and whitespace draw
+
 ## 0.2.0 - 2024-09-25
 
 - add StaticText to draw text in Rectangle area